Organizes layout component
